pyproject.toml
[project]
name = "letdem"
version = "0.1.0"
description = "letdem backend"
authors = [
    {name = "PergenDeveloper",email = "pergendeveloper@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "django (>=5.1.4,<6.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "boto3 (>=1.35.99,<2.0.0)",
    "django-storages (>=1.14.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "celery (>=5.4.0,<6.0.0)",
    "django-celery-beat (>=2.7.0,<3.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "firebase-admin (>=6.6.0,<7.0.0)",
    "django-ses (>=4.4.0,<5.0.0)",
    "django-cors-headers (>=4.6.0,<5.0.0)",
    "django-polymorphic (>=3.1.0,<4.0.0)",
    "drf-extra-fields (>=3.7.0,<4.0.0)",
    "mapbox (>=0.18.1,<0.19.0)",
    "geojson (>=3.2.0,<4.0.0)",
    "pillow (>=11.1.0,<12.0.0)",
    "django-dynamic-preferences (>=1.17.0,<2.0.0)",
    "django-rest-polymorphic (>=0.1.10,<0.2.0)",
    "django-cleanup (>=9.0.0,<10.0.0)",
    "django-waffle (>=4.2.0,<5.0.0)",
    "onesignal-python-api (>=2.0.2,<3.0.0)",
    "phonenumbers (>=9.0.2,<10.0.0)",
    "stripe (>=12.0.0,<13.0.0)",
    "python-stdnum (>=1.20,<2.0)",
    "channels-redis (>=4.3.0,<5.0.0)",
    "python-geohash (>=0.8.5,<0.9.0)",
    "setuptools (>=80.9.0,<81.0.0)",
    "channels[daphne] (>=4.2.2,<5.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest-django = "^4.10.0"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "letdem.settings.test"

